configured some aliases for swift.

+++ .zshrc
@@ -10,17 +10,25 @@ setopt cdable_vars
 
 PROMPT="%n@%m:%~ %% "
 
-alias ls="ls -F --color=auto"
-alias cp="cp -i"
-alias mv="mv -i"
-alias rm="rm -i"
+unamerslt=`uname`
+
+if [ $unamerslt = "Darwin" ]; then
+    alias ls="gls -F --color=auto"
+    alias cp="gcp -i"
+    alias mv="gmv -i"
+    alias rm="grm -i"
+fi
+if [ $unamerslt = "Linux" ]; then
+    alias ls="ls -F --color=auto"
+    alias cp="cp -i"
+    alias mv="mv -i"
+    alias rm="rm -i"
+fi
 alias start-emacs-daemon="emacs --daemon"
 alias start-emacs-client="emacsclient --create-frame &"
 alias start-emacs-client-nw="emacsclient -nw"
 alias start-emacs-client-terminal="emacsclient -nw"
 alias kill-emacs-daemon="emacsclient -e '(kill-emacs)'"
-
-unamerslt=`uname`
 
 if [ $unamerslt != "Linux" ]; then
     source $(brew --prefix nvm)/nvm.sh
